+++ wp-content/plugins/tatsu/includes/modules/icon_card.php
@@ -14,15 +14,19 @@ if( !function_exists( 'tatsu_icon_card' ) ) {
             'icon_style'        => '',
             'icon_bg'           => '',
             'icon_color'        => '',
+			'icon_hover_color'  => '',
             'size'              => 'medium',
             'line_animate'      => '0',
             'svg_icon_color'    => '',
+			'svg_icon_hover_color' => '',
 			'box_shadow'        => '', // Icon's Shadow
 			'outer_box_shadow' => '',
             'title'             => '',
             'url'               => '',
+			'new_tab'           => 0,
             'title_font'        => '',
             'title_color'       => '',
+			'title_hover_color' => '',
             'caption_font'      => '',
             'caption_color'     => '',
             'margin'            => '0 0 60px 0',
@@ -80,7 +84,8 @@ if( !function_exists( 'tatsu_icon_card' ) ) {
         $icon = !empty($icon) ? $icon : '';
         
 		$icon_label = ucfirst(str_replace(array('tatsu','icon'),' ',str_replace('-','',$icon)));
-        
+        $new_tab = $new_tab ? 'target="_blank"' : '';
+
         $classes = implode( ' ', $classes );
         ob_start();
     ?>
@@ -105,7 +110,7 @@ if( !function_exists( 'tatsu_icon_card' ) ) {
 									echo $title;
 								}else{
 							?>
-                                <a href = "<?php echo $url; ?>">
+                                <a href = "<?php echo $url; ?>" <?php echo $new_tab; ?> >
                                     <?php echo $title; ?>
                                 </a>
 							<?php } ?>
@@ -183,6 +188,7 @@ function tatsu_register_icon_card()
 							'bg_size',
 							'title',
 							'url',
+							'new_tab',
 							'content',
 						)
 					),
@@ -208,12 +214,33 @@ function tatsu_register_icon_card()
 										'type' => 'panel',
 										'title' => esc_html__('Colors', 'tatsu'),
 										'group' => array(
-											'icon_color',
-											'icon_bg',
-											'box_shadow',
-											'svg_icon_color',
-											'title_color',
-											'caption_color',
+											array(
+												'type'	=>	'tabs',
+												'style'	=>	'style1',
+												'group'	=>	array(
+													array(
+														'type'	=>	'tab',
+														'title'	=>	esc_html__('Normal', 'tatsu'),
+														'group'	=>	array(
+															'icon_color',
+															'svg_icon_color',
+															'title_color',
+															'caption_color',
+															'icon_bg',
+															'box_shadow',
+														)
+													),
+													array(
+														'type'	=>	'tab',
+														'title'	=>	esc_html__('Hover', 'tatsu'),
+														'group'	=>	array(
+															'icon_hover_color',
+															'svg_icon_hover_color',
+															'title_hover_color',
+														)
+													),
+												)
+											),
 										)
 									),
 									array(
@@ -341,7 +368,7 @@ function tatsu_register_icon_card()
 				'label' => esc_html__('Image', 'tatsu'),
 				'visible' => array('icon_type', '=', 'image'),
 				'tooltip' => '',
-				'default'	=> 'http://placehold.it/150x150',
+				'default'	=> 'http://via.placeholder.com/150x150',
 				'options' => array(
 					'size' => 'thumbnail',
 				),
@@ -459,6 +486,24 @@ function tatsu_register_icon_card()
 				)
 			),
 			array(
+				'att_name' => 'icon_hover_color',
+				'type' => 'color',
+				'visible'	=> array('icon_type', '=', 'icon'),
+				'options' => array(
+					'gradient' => true,
+				),
+				'label' => esc_html__('Icon Hover Color', 'tatsu'),
+				'default' => '',
+				'tooltip' => '',
+				'css'	  => true,
+				'selectors'	=> array(
+					'.tatsu-{UUID} .tatsu-icon:hover'	=> array(
+						'property'		=> 'color',
+						'when'			=> array('icon_type', '=', 'icon')
+					)
+				)
+			),
+			array(
 				'att_name'	=> 'svg_icon_color',
 				'type' => 'color',
 				'visible'	=> array('icon_type', '=', 'svg'),
@@ -468,6 +513,21 @@ function tatsu_register_icon_card()
 				'css'	  => true,
 				'selectors'	=> array(
 					'.tatsu-{UUID} .tatsu-icon_card-icon svg' => array(
+						'property'		=> 'color',
+						'when'			=> array('icon_type', '=', 'svg')
+					)
+				)
+			),
+			array(
+				'att_name'	=> 'svg_icon_hover_color',
+				'type' => 'color',
+				'visible'	=> array('icon_type', '=', 'svg'),
+				'label' => esc_html__('Icon Color', 'tatsu'),
+				'default' => '',
+				'tooltip' => '',
+				'css'	  => true,
+				'selectors'	=> array(
+					'.tatsu-{UUID} .tatsu-icon_card-icon svg:hover' => array(
 						'property'		=> 'color',
 						'when'			=> array('icon_type', '=', 'svg')
 					)
@@ -520,6 +580,13 @@ function tatsu_register_icon_card()
 				'default' => '',
 				'tooltip' => ''
 			),
+			array(
+				'att_name'	 => 'new_tab',
+				'type' => 'switch',
+				'label' => esc_html__('Open Link in New Tab', 'tatsu'),
+				'default' => 0,
+				'tooltip' => ''
+			),
 			function_exists('typehub_get_exposed_selectors') ? array(
 				'att_name'	=> 'title_font',
 				'type'		=> 'select',
@@ -538,7 +605,23 @@ function tatsu_register_icon_card()
 				),
 				'css'	=> true,
 				'selectors'	=> array(
-					'.tatsu-{UUID} .tatsu-icon_card-title' => array(
+					'.tatsu-{UUID} .tatsu-icon_card-title, .tatsu-{UUID} .tatsu-icon_card-title a' => array(
+						'property'		=> 'color',
+					)
+				)
+			),
+			array(
+				'att_name' => 'title_hover_color',
+				'type' => 'color',
+				'label' => esc_html__('Title Hover Color', 'tatsu'),
+				'default' => '',
+				'tooltip' => '',
+				'options'	=> array(
+					'gradient'	=> true
+				),
+				'css'	=> true,
+				'selectors'	=> array(
+					'.tatsu-{UUID} .tatsu-icon_card-title:hover, .tatsu-{UUID} .tatsu-icon_card-title a:hover' => array(
 						'property'		=> 'color',
 					)
 				)