+++ wp-content/plugins/tatsu/includes/modules/icon_group.php
@@ -85,7 +85,8 @@ function tatsu_register_icon_group()
 		'is_js_dependant' => false,
 		'child_module' => 'tatsu_icon',
 		'type' => 'multi',
-		'is_built_in' => true,
+		'allowed_sub_modules' => array('tatsu_icon'),
+		'is_built_in' => false,
 		'group_atts' => array(
 			array(
 				'type'		=> 'tabs',