+++ wp-content/plugins/exponent-modules/changelog.md
@@ -1,3 +1,7 @@
+v2.1.5 (9 February 2022)
+======
+* Added: Custom text, color and background color of success message in newsletter
+
 v2.1.4
 ======
 * Updated: pot file