+++ wp-content/plugins/tatsu/changelog.md
@@ -1,3 +1,63 @@
+v3.3.14 (2 June 2022)
+=====
+* Added: Module Categories
+* Fixed: TGMPA notice dismiss
+* Fixed: Placeholder URL
+* Fixed: Hamburger Menu CSS issue on mobile
+* Fixed: Interactive Box CSS issue
+* Fixed: Template Import issue
+* Fixed: Wrong conditional tag
+
+v3.3.13 (30 March 2022)
+=====
+* Fixed: Form Submission email
+* Fixed: Typehub Custom font upload conditional logic
+* Added: Tooltip for penultimate settings
+
+v3.3.12 (16 March 2022)
+=====
+* Fixed: Tatsu header logo for non SVG images
+* Fixed: Typehub download Google Fonts
+* Fixed: Typehub load Google Fonts from local server
+* Fixed: Typehub and Colorhub activation and deactivation action
+* Fixed: Php deprecation
+* Fixed: Pro modules list
+* Fixed: Typehub file upload check for add custom font 
+* Fixed: Tatsu header mobile sliding menu styles
+
+v3.3.11 (9 March 2022)
+=====
+* Added: Sidebar in Tatsu Welcome page.
+* Fixed: Spyro Notice will be visible on only Tatsu welcome page
+* Fixed: Prevent Tatsu admin dashboard files load on other admin pages.
+* Fixed: Separate standalone tatsu theme css
+* Fixed: Standalone Tatsu header and footer builder template
+* Fixed: Standalone Tatsu replace theme header and footer with active Tatsu header and footer respectively
+
+v3.3.10 (24 February 2022)
+=====
+* Fixed: Tatsu header logo for SVG images
+
+v3.3.9 (22 February 2022)
+=====
+* Added: Multi Purpose Card Hover Color
+* Added: Multi Purpose Card Link Open in New Tab
+* Added: Custom Image can be set in Icon Module and Icon Group Module
+* Fixed: Autoplay audio with video for vimeo and self-hosted video
+* Fixed: Image carousel image alignment on lazyload option
+* Fixed: Image carousel module ribbon and ribbon center type layout
+* Fixed: GIF image display in single image module
+* Fixed: Border on Single Image module
+* Fixed: Mobile submenu
+* Fixed: Quoted text in Animated Heading
+* Fixed: Icon Module CSS issues
+
+v3.3.8
+=====
+* Fixed: Spyro form mailchimp integration
+* Fixed: Save Spyro form multiple checkbox values
+* Fixed: Spyro Forms required field handling for checkbox
+
 v3.3.7
 =====
 * Fixed: Revslider background video with tatsu youtube video